refactor: moved use-usage-data into use-rules-engine and removed console logs
