+++ heat-stack/app/utils/hooks/use-rules-engine.ts
@@ -2,12 +2,15 @@ import { useEffect, useRef, useState } from 'react'
 import {
 	buildCurrentMapOfUsageData,
 	buildCurrentUsageData,
+	hasDataProperty,
 	objectToString,
 } from '#app/utils/index.ts'
+import { type EnergyUsageSubmitActionData } from '#types/single-form.ts'
 import {
 	type UsageDataSchema,
 	type BillingRecordsSchema,
 } from '#types/types.ts'
+import { reviver } from '../data-parser'
 
 // importRulesEngine exists so the rules-engine.ts types can be extracted without having to manually create a types.ts file.
 const importRulesEngine = () => import('#app/utils/rules-engine.ts')
@@ -19,22 +22,18 @@ export type RecalculateFunction = NonNullable<
 >
 
 /**
- * Custom React hook for interacting with a Pyodide-backed rules engine.
- *
- * Exposes functions:
- * 1. `lazyLoadRulesEngine`: A fire-and-forget function that lazily loads and initializes the rules engine, but does **not** return a promise
- *    or wait for completion.
- *
- * 2. `recalculateFromBillingRecordsChange`: A function that is only available **after** the rules engine has been initialized. If not yet initialized,
- *    this will be `null`.
+ * Custom React hook for interacting with a Pyodide-backed rules engine and keeps track of usage data.
  *
  * Usage pattern:
  * - Call lazyLoadRulesEngine() and wait for `recalculateFromBillingRecordsChange` to transition from null to a function.
  *
  */
-export const useRulesEngine = () => {
+export const useRulesEngine = (
+	actionData: EnergyUsageSubmitActionData | undefined,
+) => {
 	const [isInitialized, setIsInitialized] = useState(false)
 	const rulesEngineRef = useRef<RulesEngineType | null>(null)
+	const [usageData, setUsageData] = useState<UsageDataSchema | undefined>()
 
 	const lazyLoadRulesEngine = () => {
 		if (isInitialized) {
@@ -58,9 +57,6 @@ export const useRulesEngine = () => {
 		convertedDatesTIWD: any,
 		state_id: any,
 		county_id: any,
-		setUsageData: React.Dispatch<
-			React.SetStateAction<UsageDataSchema | undefined>
-		>,
 	) => {
 		const rulesEngine = rulesEngineRef.current
 
@@ -108,6 +104,57 @@ export const useRulesEngine = () => {
 		})
 	}
 
+	/**
+	 * Toggles the selection state of a table row corresponding to a specific billing period,
+	 * and updates the overall energy usage calculation.
+	 *
+	 * @param index - The index of the billing period row whose checkbox was clicked.
+	 */
+	const toggleBillingPeriod = (index: number) => {
+		if (!usageData) {
+			// if this function was called and there is no usage data, something went wrong
+			throw new Error('usageData was not found')
+		}
+
+		if (!actionData) {
+			// if this function was called and there is no actionData, something went wrong
+			throw new Error('actionData was not found')
+		}
+
+		if (!recalculateFromBillingRecordsChange) {
+			// if this function was called and there is no recalculateFromBillingRecordsChange then pyodide was not initialized and something went wrong
+			throw new Error('recalculateFromBillingRecordsChange was not found')
+		}
+
+		const newRecords = structuredClone(usageData.processed_energy_bills)
+		const period = newRecords[index]
+
+		if (!period) {
+			throw new Error(`No period with row index ${index} was found`)
+		}
+
+		const currentOverride = period.inclusion_override
+
+		// Toggle 'inclusion_override'
+		period.inclusion_override = !currentOverride
+
+		newRecords[index] = { ...period }
+
+		recalculateFromBillingRecordsChange(
+			// TODO: I think we should be passing usageDataAsMap BUT there will be an error thrown because
+			// usageDataAsMap.get('balance_point_graph') changes from a map to a map of proxies resulting
+			// in an error thrown by structuredClone inside of buildCurrentMapOfUsageData.
+			// Previous implementation passed actionData.data to buildCurrentMapOfUsageData which feels wrong but works.
+			JSON.parse(actionData.data, reviver) as Map<any, any> | undefined,
+			newRecords,
+			actionData.parsedAndValidatedFormSchema,
+			actionData.convertedDatesTIWD,
+			actionData.state_id,
+			actionData.county_id,
+		)
+	}
+
+	// shutdown pyodide when component unmounts
 	useEffect(() => {
 		return () => {
 			// Memory cleanup of pyodide fn's when component unmounts
@@ -118,8 +165,22 @@ export const useRulesEngine = () => {
 		}
 	}, [])
 
+	// reset usage data as a result of user submitting a new bill
+	useEffect(() => {
+		if (actionData !== undefined && hasDataProperty(actionData)) {
+			const usageDataAsMap = JSON.parse(actionData.data, reviver) as Map<
+				any,
+				any
+			>
+			const newUsageData = buildCurrentUsageData(usageDataAsMap)
+			setUsageData(newUsageData)
+		}
+	}, [actionData])
+
 	return {
 		lazyLoadRulesEngine,
+		toggleBillingPeriod,
+		usageData,
 		recalculateFromBillingRecordsChange: isInitialized
 			? recalculateFromBillingRecordsChange
 			: null,
